Add basic parent Boundary object
The Boundary object will be the superclass to the
FFRBoundary and InteriorBoundary. The only properties
which both share are the bound value (either an x or y value),
and the count of total photon crossings of that boundary.
The increment function is provided to avoid verbosity.
Make this a handle class (variables by reference, not by value)
so we don't have to create a new Boundary each time its
properties change (i.e. a photon crossing).