Add the InteriorBoundary
The FFRBoundary inherits from Boundary. Its only additional
property is its "range". The range is a value to be added
and subtracted to the bound value to get the upperBound and
lowerBound, respectively. We need to give interior bounds
a thickness because that's the only way to check if a photon
has crossed the bound when we have more than one interior
bound (a photon just being above or below the interior bound
doesn't indicate whether a crossing has happened, nor even
when taking into account the direction of the photon). We
use Java's HashSet to store the IDs of photon which have
crossed the Bound to track repeats.