Reduce checkIfReflected() runtime with findCurrentQuadrant()
Instead of iterating through each fiber to check
if a Photon has reflected, we can just iterate through
the fiber's in the Quadrant which the Photon is in (the
Photon can't have reflected off fibers outside its Quadrant,
by definition). This is a matter of iterating through each
Quadrant and comparing the photon's coordinate with the
boundary coordinates of each Quadrant (different and
independent from the Boundary objects).