Rewrite bound crossing detection logic
See comments under checkIfAtInteriorBound() for a detailed explanation.
In short, however, we keep track of the current and previous FFR layer
which the photon is in. Each time the photon is moved, we test if the
current layer is different that the previous layer. If so, the photon
has crossed an interior boundary. The actual code here is very simple,
just a simple isequal(curr, prev) test.

Comment out previous attempt.