Add variables and functions for detecting crossings
Store the current and previous FFR layers as properties, and set
the current layer upon instantiation. We then need a function to
find the current FFR layer (plain iteration) so we can update it.
We also need to find the actual boundary that was crossed between
two FFR layers (the shared boundary between the two layers).