Track and update the previous and current FFRLayer
As part of the new method to detect interior bound crossings, we need
to track the previous and current FFRLayer which the photon is in. Use
properties for these values and reset the current layer before each new
incident photon is ray traced. Store the previous layer from the current
layer before the photon is moved. Update the current layer with the
findCurrentFFRLayer() function.