Make Config FFR-agnostic w/ specific config structs
Make each build*() function independent of the actual FFR (this is
a necessary feature if Config is to be static and used in multiple
FFR configs). Give each FFR its own config struct with user-defined
values so the user doesn't have to edit the respective config_*.m file,
which is risky.