Vectorize checkIfReflected() and distanceToFiber()
Instead of checking for fiber reflections one fiber at a time, use
logical indexing in combination with vectorized reflection radii
and photon distances to extract the reflected fiber (if it exists).
The new distancesToFiber() function "vectorially" calculates the
distance of the photon to all the fibers. The new checkIfReflected()
function vectorially compares the distances with the reflection radii.