Increase photonsPaths and store every 100th coord
Initialize the photonPaths array which stores the photon coordinates
which will be plotted to 10,000,000 from 1,000,000. This helps guarantee
the simulator won't run out of space when a simulation results in more
photon coordinates than expected.

Store every 100th photon coordinate instead of every photon coordinate.
This allows 100 times more coordinates to be stored and does not affect
the visualization.