Add functions to save data, figure, and config
Add photonsInToCSV() to write the FFRLayer.nPhotonsIn number to a
csv representing the nPhotonsIn per FFRLayer, counting down from
nLayers to 1. This is a nLayers x 2 csv file and can be parsed
or opened in a spreadsheet for analysis.

Add configToMAT() to write the ffrConfig struct to a .mat file as
a backup in case we need to refer to some configuration value for
a given simulation. The complexity of the ffrConfig struct made it
easiest to write to .mat instead of .txt or something else.

Add figureToSVG() to write the visualization to an svg file. SVG
can be transformed to PNG, PDF, etc, and is not too big as a file.
It can also be easily integrated into a LaTeX document.

To facilitate these functions, we need to add the 'model' value to
the FFR config. This is just a string like "3M9210" or "3M1870".