Redo some plotting to remove text labels
The text labels were not placed well and I removed them and will
deal with them later. To make plotting optional, wrap all the plotting
statements in an if..end block toggled by the `plotbool` variable.

This includes at attempt at a plotFFRBounds() function which plots the
bounds as line segments instead of infinite lines. This can be done later.