Add script to calculate number of remaining viruses
See the docstring. This script basically gives us one data point in
our "viral load vs. number of photons" graph by calculating the
number of viruses remaining for a given number of photons available
for decontamination in the outer layer. It uses the photon-data.csv
and the virus-data.csv files as inputs, but this will likely be moved
to a different directory.