Match changes made in 852c552
I commited changes to RayTracer.m and RayTracer.org in 852c552.
This was a mistake, as I should have just commited the RayTracer.m
changes so that I could cherry-pick them without the org file.
So I just copied the actual literate branch RayTracer.m file to this
one, and am commiting the changes normally.